+++ packages/elements-web-components/src/__stories__/index.ts
@@ -1,0 +1,2 @@
+export * from './SchemaViewer';
+export * from './StoplightProject';