+++ packages/elements-web-components/src/index.ts
@@ -1,0 +1,4 @@
+import { SchemaViewerComponentElement, StoplightProjectElement } from './components';
+
+window.customElements.define('stoplight-component-schema-viewer', SchemaViewerComponentElement);
+window.customElements.define('stoplight-project', StoplightProjectElement);