+++ packages/web-components/src/index.ts
@@ -1,3 +1,0 @@
-import { SchemaViewerComponentElement } from './components';
-
-window.customElements.define('stoplight-component-schema-viewer', SchemaViewerComponentElement);