+++ packages/elements-web-components/CHANGELOG.md
@@ -3,6 +3,39 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [6.0.0-beta.128](https://github.com/stoplightio/elements/compare/v6.0.0-beta.127...v6.0.0-beta.128) (2020-08-21)
+
+
+### Bug Fixes
+
+* remove cycles before running openapi-sampler ([#543](https://github.com/stoplightio/elements/issues/543)) ([747e6ae](https://github.com/stoplightio/elements/commit/747e6ae0bc68fb623e375d060d45da4ae8104ab0))
+
+
+
+
+
+# [6.0.0-beta.127](https://github.com/stoplightio/elements/compare/v6.0.0-beta.126...v6.0.0-beta.127) (2020-08-20)
+
+
+### Features
+
+* stoplight project container ([#529](https://github.com/stoplightio/elements/issues/529)) ([07d0444](https://github.com/stoplightio/elements/commit/07d0444ec3844ba305e8512ced90454b035eaeb3)), closes [#539](https://github.com/stoplightio/elements/issues/539)
+
+
+
+
+
+# [6.0.0-beta.123](https://github.com/stoplightio/elements/compare/v6.0.0-beta.122...v6.0.0-beta.123) (2020-08-11)
+
+
+### Features
+
+* **web-components:** publish as @stoplight/elements-web-components ([#527](https://github.com/stoplightio/elements/issues/527)) ([aebc77a](https://github.com/stoplightio/elements/commit/aebc77a4c70d8af9c9dd4b80496855ce56ec03c6))
+
+
+
+
+
 # [6.0.0-beta.121](https://github.com/stoplightio/elements/compare/v6.0.0-beta.119...v6.0.0-beta.121) (2020-08-07)